added alert to confirm the removal

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -35,7 +35,25 @@ class Favorites extends Component {
         {
           text: 'Delete',
           type: 'delete',
-          onPress: () => this.props.deleteFav(item.id)
+          onPress: () => {
+            Alert.alert(
+              'Remove Favourite',
+              `Are you sure you want to remove ${item.name} from your favourites?`,
+              [
+                {
+                  text: 'Cancel', 
+                  onPress: () => console.log('Not removed'),
+                  style: 'cancel'
+                },
+                {
+                  text: 'Ok', 
+                  onPress: () => this.props.deleteFav(item.id),
+                  style: 'default'
+                }
+              ],
+              { cancelable: false }
+            )
+          }
         }
       ];
       return (
